+++ security/oc_tls.c
@@ -460,7 +460,7 @@ oc_tls_shutdown(void)
 #endif /* OC_TCP */
   }
 #endif /* OC_CLIENT */
-  int device;
+  size_t device;
   for (device = 0; device < oc_core_get_num_devices(); device++) {
     mbedtls_ssl_config_free(&server_conf[device]);
 #ifdef OC_TCP
@@ -515,7 +515,7 @@ oc_tls_init_context(void)
                                &ctr_drbg_ctx) != 0) {
     goto dtls_init_err;
   }
-  int i;
+  size_t i;
 
 #ifdef OC_TCP
 #define mbedtls_config_tls(func_name, conf, index, ...)                        \
@@ -625,7 +625,7 @@ dtls_init_err:
 }
 
 int
-oc_tls_update_psk_identity(int device)
+oc_tls_update_psk_identity(size_t device)
 {
   oc_uuid_t *device_id = oc_core_get_device_id(device);
   if (!device_id) {
